add condition preventing errors

+++ script/patch_arduino.py
@@ -7,8 +7,11 @@ name = env.get( "PIOENV" )#print(FRAMEWORK_DIR)
 board = env.BoardConfig().get_brief_data()['id']
 dir = env.GetLaunchDir()
 
-source = dir + "/boards/"+ board +".pins.h"
-pathlib.Path(framework_dir + "/variants/"+ board).mkdir(parents=True, exist_ok=True)
-target = framework_dir + "/variants/"+ board +"/pins_arduino.h"
-print( "Copy file : " + source + " -> " + target )
-shutil.copyfile( source, target )
+isCustomPins = pathlib.Path(dir + "/boards/"+ board +".pins.h")
+if isCustomPins.is_file():
+    # check if file exists to prevent errors
+    source = dir + "/boards/"+ board +".pins.h"
+    pathlib.Path(framework_dir + "/variants/"+ board).mkdir(parents=True, exist_ok=True)
+    target = framework_dir + "/variants/"+ board +"/pins_arduino.h"
+    print( "Copy file : " + source + " -> " + target )
+    shutil.copyfile( source, target )
